working fft

+++ fft/twiddle.h
@@ -4,5 +4,5 @@ struct cnum {
 	float real;
 	float im;
 };
-struct cnum twiddle(float real, float im, uint16_t k, uint16_t b);
+struct cnum twiddle(float real, float im, unsigned int k, unsigned int b);
 #endif /* __TWIDDLE_H__ */
